+++ config/other/constants.py
@@ -1,7 +1,5 @@
 import os
 
-from utility import Settings
-
 SONARR_URL = os.getenv('SONARR_URL') # Indirizzo ip + porta di sonarr
 API_KEY = os.getenv('API_KEY') # Chiave api di sonarr
 VERSION = os.getenv('VERSION') # versione
