+++ config/other/texts.py
